feat: Improve local keyring backend selection for storing PAT and implement automatic token refresh
The code changes modify the `ArtifactsKeyringBackend` class to improve the selection of the local keyring backend for storing the Personal Access Token (PAT). The changes introduce a new private attribute `__LOCAL_BACKEND` to prevent infinite looping when falling back to the local backend. Additionally, the `_LOCAL_BACKEND` property method is updated to return the local backend if it has already been set, and if not, set it and return it.

Expired tokens now get automatically refreshed. This is achieved by checking if authentication with the stored token is possible and if not a the `_PROVIDER` will fetch a new one. Changes implemented in `get_credential` method